Add OQtrl_params.py and make changes to OQtrl.py
